15/11/2022

+++ app/src/main/java/com/ptbas/controlcenter/adapter/CashOutManagementAdapter.java
@@ -204,7 +204,7 @@ public class CashOutManagementAdapter extends RecyclerView.Adapter<CashOutManage
                                     }
 
                                     //Toast.makeText(context, "buy "+matBuyPrice+" - unit"+totalUnit, Toast.LENGTH_LONG).show();
-                                    double totalIDR = matBuyPrice * Double.parseDouble(df.format(totalUnit));
+                                    double totalIDR = matBuyPrice * Float.parseFloat(df.format(totalUnit));
                                     coTotal = "IDR " + currencyFormat(df.format(Float.valueOf(String.valueOf(totalIDR))));
                                     tvTotalDue.setText(coTotal);
                                 }
